テーマの読み込みをインラインscriptタグに移動
